+++ src/__tests__/index-test.js
@@ -40,9 +40,7 @@ Array [
   "DatePickerInput",
   "DatePickerSkeleton",
   "Dropdown",
-  "DropdownItem",
   "DropdownSkeleton",
-  "DropdownV2",
   "ExpandableTile",
   "FileUploader",
   "FileUploaderButton",
