+++ src/components/Dropdown/Dropdown-test.js
@@ -6,233 +6,188 @@
  */
 
 import React from 'react';
-import { iconCaretDown } from 'carbon-icons';
+import { mount, shallow } from 'enzyme';
+import {
+  assertMenuOpen,
+  assertMenuClosed,
+  findMenuItemNode,
+  openMenu,
+  generateItems,
+  generateGenericItem,
+} from '../ListBox/test-helpers';
 import Dropdown from '../Dropdown';
-import DropdownItem from '../DropdownItem';
-import Icon from '../Icon';
-import ClickListener from '../../internal/ClickListener';
-import { shallow, mount } from 'enzyme';
+import DropdownSkeleton from '../Dropdown/Dropdown.Skeleton';
 
-describeBreakingChangesXFeatures('Dropdown', () => {
-  describe('Renders as expected', () => {
-    const wrapper = shallow(
-      <Dropdown className="extra-class" defaultText="Choose something.." />
-    );
-    const dropdownWrapper = wrapper.childAt(0);
-    const mounted = mount(
-      <Dropdown className="extra-class" defaultText="Choose something.." />
-    );
+describe('Dropdown', () => {
+  let mockProps;
+  beforeEach(() => {
+    mockProps = {
+      items: generateItems(5, generateGenericItem),
+      onChange: jest.fn(),
+      label: 'input',
+      placeholder: 'Filter...',
+      type: 'default',
+    };
+  });
 
-    it('renders a dropdown', () => {
-      expect(wrapper.length).toEqual(1);
+  it('should render', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should initially render with the menu not open', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    assertMenuClosed(wrapper);
+  });
+
+  it('should let the user open the menu by clicking on the control', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    openMenu(wrapper);
+    assertMenuOpen(wrapper, mockProps);
+  });
+
+  it('should render with strings as items', () => {
+    const wrapper = mount(<Dropdown {...mockProps} items={['zar', 'doz']} />);
+    openMenu(wrapper);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render custom item components', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    wrapper.setProps({
+      itemToElement: item => <div className="mock-item">{item.label}</div>,
+    });
+    openMenu(wrapper);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  describe('title', () => {
+    const wrapper = mount(<Dropdown titleText="Email Input" {...mockProps} />);
+    const renderedLabel = wrapper.find('label');
+
+    it('renders a title', () => {
+      expect(renderedLabel.length).toBe(1);
     });
 
     it('has the expected classes', () => {
-      expect(dropdownWrapper.hasClass('bx--dropdown')).toEqual(true);
+      expect(renderedLabel.hasClass('bx--label')).toEqual(true);
     });
 
-    it('has the expected classes when disabled', () => {
-      const wrapper = shallow(
-        <Dropdown defaultText="Choose something.." disabled />
-      ).childAt(0);
-
-      expect(wrapper.hasClass('bx--dropdown--disabled')).toEqual(true);
-    });
-
-    it('has the expected classes when light is true', () => {
-      const wrapper = shallow(
-        <Dropdown defaultText="Choose something.." light />
-      ).childAt(0);
-
-      expect(wrapper.hasClass('bx--dropdown--light')).toEqual(true);
-    });
-
-    it('should add extra classes that are passed via className', () => {
-      expect(dropdownWrapper.hasClass('extra-class')).toEqual(true);
-    });
-
-    it('should render children as expected', () => {
-      const dropdown = shallow(
-        <Dropdown>
-          <div className="test-child" />
-          <div className="test-child" />
-        </Dropdown>
-      );
-      expect(dropdown.find('.test-child').length).toEqual(2);
-    });
-
-    it('should handle null children', () => {
-      const dropdown = shallow(
-        <Dropdown>
-          {null}
-          <div className="test-child" />
-          {null}
-        </Dropdown>
-      );
-      expect(dropdown.find('.test-child').length).toEqual(1);
-    });
-
-    it('should use correct icon', () => {
-      const icon = mounted.find(Icon);
-      expect(icon.props().icon).toEqual(iconCaretDown);
-    });
-
-    it('has the expected default iconDescription', () => {
-      expect(mounted.props().iconDescription).toEqual('open list of options');
-    });
-
-    it('adds new iconDescription when passed via props', () => {
-      mounted.setProps({ iconDescription: 'new description' });
-      expect(mounted.props().iconDescription).toEqual('new description');
-    });
-
-    it('should have iconDescription match Icon component description prop', () => {
-      const matches =
-        mounted.props().iconDescription ===
-        mounted.find(Icon).props().description;
-      expect(matches).toEqual(true);
-    });
-
-    it('should start with the selected text over the default text when present in props', () => {
-      const dropdown = shallow(
-        <Dropdown defaultText="Choose something..." selectedText="Value">
-          <DropdownItem itemText="Value" value="Value" />
-        </Dropdown>
-      );
-      expect(dropdown.state().selectedText).toEqual('Value');
-    });
-
-    it('should select default text when provided selected text does not match any children', () => {
-      const dropdown = shallow(
-        <Dropdown defaultText="Choose something..." selectedText="NotValue">
-          <DropdownItem itemText="Value" value="Value" />
-        </Dropdown>
-      );
-      expect(dropdown.state().selectedText).toEqual('Choose something...');
+    it('should set title as expected', () => {
+      expect(renderedLabel.text()).toEqual('Email Input');
     });
   });
 
-  describe('events', () => {
-    const onClick = jest.fn();
-    const onOpen = jest.fn();
-    const onClose = jest.fn();
-    const handleKeydown = jest.fn();
+  describe('helper', () => {
+    it('renders a helper', () => {
+      const wrapper = mount(
+        <Dropdown helperText="Email Input" {...mockProps} />
+      );
+      const renderedHelper = wrapper.find('.bx--form__helper-text');
+      expect(renderedHelper.length).toEqual(1);
+    });
 
-    const wrapper = mount(
-      <Dropdown onClick={onClick} onOpen={onOpen} onClose={onClose}>
-        <DropdownItem
-          className="test-child"
-          itemText="test-child"
-          value="test-child"
+    it('renders children as expected', () => {
+      const wrapper = mount(
+        <Dropdown
+          helperText={
+            <span>
+              This helper text has <a href="/">a link</a>.
+            </span>
+          }
+          {...mockProps}
         />
-      </Dropdown>
-    );
-
-    const dropdown = wrapper.find('.bx--dropdown');
-    const child = wrapper.find('.test-child');
-
-    it('should add the open dropdown class on click', () => {
-      dropdown.simulate('click');
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(true);
+      );
+      const renderedHelper = wrapper.find('.bx--form__helper-text');
+      expect(renderedHelper.props().children).toEqual(
+        <span>
+          This helper text has <a href="/">a link</a>.
+        </span>
+      );
     });
 
-    it('should toggle the open dropdown class on Enter', () => {
-      wrapper.setState({ open: false });
-      dropdown.simulate('keypress', { which: 13 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(true);
-      dropdown.simulate('keypress', { which: 13 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(false);
+    it('should set helper text as expected', () => {
+      const wrapper = mount(<Dropdown {...mockProps} />);
+      wrapper.setProps({ helperText: 'Helper text' });
+      const renderedHelper = wrapper.find('.bx--form__helper-text');
+      expect(renderedHelper.text()).toEqual('Helper text');
     });
+  });
 
-    it('should toggle the open dropdown class on Space', () => {
-      wrapper.setState({ open: false });
-      dropdown.simulate('keypress', { which: 32 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(true);
-      dropdown.simulate('keypress', { which: 32 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(false);
+  it('should render custom components components', () => {
+    const DropdownItem = jest.fn(() => 'Dropdown item');
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    wrapper.setProps({
+      itemToElement: DropdownItem,
     });
+    openMenu(wrapper);
+    expect(wrapper).toMatchSnapshot();
+  });
 
-    it('should close when ESC is pressed', () => {
+  it('should specify light version as expected', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    expect(wrapper.props().light).toEqual(false);
+    wrapper.setProps({ light: true });
+    expect(wrapper.props().light).toEqual(true);
+  });
+
+  it('should let the user select an option by clicking on the option node', () => {
+    const wrapper = mount(<Dropdown {...mockProps} />);
+    openMenu(wrapper);
+    findMenuItemNode(wrapper, 0).simulate('click');
+    expect(mockProps.onChange).toHaveBeenCalledTimes(1);
+    expect(mockProps.onChange).toHaveBeenCalledWith({
+      selectedItem: mockProps.items[0],
+    });
+    assertMenuClosed(wrapper);
+
+    mockProps.onChange.mockClear();
+
+    openMenu(wrapper);
+    findMenuItemNode(wrapper, 1).simulate('click');
+    expect(mockProps.onChange).toHaveBeenCalledTimes(1);
+    expect(mockProps.onChange).toHaveBeenCalledWith({
+      selectedItem: mockProps.items[1],
+    });
+  });
+
+  describe('should display initially selected item found in `initialSelectedItem`', () => {
+    it('using an object type for the `initialSelectedItem` prop', () => {
       const wrapper = mount(
-        <Dropdown handleKeydown={handleKeydown}>
-          <DropdownItem
-            className="test-child"
-            itemText="test-child"
-            value="test-child"
-          />
-        </Dropdown>
+        <Dropdown {...mockProps} initialSelectedItem={mockProps.items[0]} />
       );
 
-      wrapper.setState({ open: true });
-      wrapper.mount();
-      dropdown.simulate('keypress', { which: 27 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(false);
-      dropdown.simulate('keypress', { which: 27 });
-      expect(
-        wrapper.find('.bx--dropdown').hasClass('bx--dropdown--open')
-      ).toEqual(false);
-      wrapper.unmount();
-    });
-
-    it('should update data value state when child item is clicked', () => {
-      child.last().simulate('click');
-      expect(wrapper.find('.bx--dropdown').props().value).toEqual('test-child');
-    });
-
-    it('should update selected text when child item is clicked', () => {
-      child.last().simulate('click');
-      expect(wrapper.state().selectedText).toEqual('test-child');
-    });
-
-    it('should close dropdown on click outside', () => {
-      wrapper.setState({ open: true });
-      const listener = wrapper.find(ClickListener);
-      listener.props().onClickOutside();
-      expect(wrapper.state().open).toBe(false);
-    });
-
-    it('fires open and close callbacks when the dropdown is clicked', () => {
-      onOpen.mockReset();
-      onClose.mockReset();
-      dropdown.simulate('click');
-      expect(onOpen).toHaveBeenCalled();
-      expect(onClose).not.toHaveBeenCalled();
-      dropdown.simulate('click');
-      expect(onOpen).toHaveBeenCalled();
-      expect(onClose).toHaveBeenCalled();
-    });
-
-    it('should not open when disabled', () => {
-      const wrapper = mount(
-        <Dropdown onClick={onClick} disabled>
-          <DropdownItem
-            className="test-child"
-            itemText="test-child"
-            value="test-child"
-          />
-        </Dropdown>
+      expect(wrapper.find('span.bx--list-box__label').text()).toEqual(
+        mockProps.items[0].label
       );
-      const dropdown = wrapper.find('.bx--dropdown--disabled');
+    });
 
-      dropdown.simulate('click');
-      expect(dropdown.hasClass('bx--dropdown--open')).toEqual(false);
-      dropdown.simulate('keypress', { which: 13 });
-      expect(dropdown.hasClass('bx--dropdown--open')).toEqual(false);
-      dropdown.simulate('keypress', { which: 32 });
-      expect(dropdown.hasClass('bx--dropdown--open')).toEqual(false);
-      expect(wrapper.state().open).toBe(false);
+    it('using a string type for the `initialSelectedItem` prop', () => {
+      // Replace the 'items' property in mockProps with a list of strings
+      mockProps = {
+        ...mockProps,
+        items: ['1', '2', '3'],
+      };
+
+      const wrapper = mount(
+        <Dropdown {...mockProps} initialSelectedItem={mockProps.items[1]} />
+      );
+
+      expect(wrapper.find('span.bx--list-box__label').text()).toEqual(
+        mockProps.items[1]
+      );
+    });
+  });
+});
+
+describe('DropdownSkeleton', () => {
+  describe('Renders as expected', () => {
+    const wrapper = shallow(<DropdownSkeleton inline />);
+
+    it('Has the expected classes', () => {
+      expect(wrapper.hasClass('bx--skeleton')).toEqual(true);
+      expect(wrapper.hasClass('bx--dropdown-v2')).toEqual(true);
+      expect(wrapper.hasClass('bx--list-box--inline')).toEqual(true);
     });
   });
 });
