+++ src/components/Dropdown/index.js
@@ -5,4 +5,5 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+export * from './Dropdown.Skeleton';
 export default from './Dropdown';
